move: perview-link.ts to tooltip/index.ts
