+++ Makefile
@@ -45,7 +45,7 @@ index:
 	sed -i "s/src=\"\/oembed\./src=\".\/oembed./g" dist/index.html
 	sed -i "s/src=\"\/format\./src=\".\/format./g" dist/index.html
 	sed -i "s/src=\"\/preview\-lia\./src=\".\/preview-lia./g" dist/index.html
-	sed -i "s/src=\"\/preview\-link\./src=\".\/preview-link./g" dist/index.html
+	sed -i "s/src=\"\/tooltip\./src=\".\/tooltip./g" dist/index.html
 	sed -i "s/src:local(\"\")/src:local(\"\.\")/g" dist/index.*.css
 	sed -i "s/url(\//url(/g" dist/index.*.css
