+++ src/entry/app/index.html
@@ -15,7 +15,7 @@
   <script defer type="module" src="../../typescript/webcomponents/embed/index.ts"></script>
   <script defer type="module" src="../../typescript/webcomponents/chart.ts"></script>
   <script defer type="module" src="../../typescript/webcomponents/preview-lia.ts"></script>
-  <script defer type="module" src="../../typescript/webcomponents/preview-link.ts"></script>
+  <script defer type="module" src="../../typescript/webcomponents/tooltip/index.ts"></script>
   <script defer type="module" src="../../typescript/webcomponents/format.ts"></script>
   <script defer type="module" src="../../typescript/webcomponents/terminal.ts"></script>
   <script>
