+++ src/entry/webcomponent/index.ts
@@ -9,10 +9,10 @@ import '../../scss/main.scss'
 
 import '../../typescript/webcomponents/editor'
 import '../../typescript/webcomponents/formula'
-import '../../typescript/webcomponents/embed'
+import '../../typescript/webcomponents/embed/index'
 import '../../typescript/webcomponents/chart'
 import '../../typescript/webcomponents/preview-lia'
-import '../../typescript/webcomponents/preview-link'
+import '../../typescript/webcomponents/tooltip/index'
 import '../../typescript/webcomponents/format'
 
 customElements.define(
